split up manipulate_local
