+++ tests/acceptance/manipulate_local.py
@@ -1,185 +1,13 @@
-__author__ = 'himanshu'
-from decorator import decorator
-import unittest
 from nose import with_setup
-import os
-import json
-import shutil
-import requests
-import time
-from osfoffline.polling_osf_manager.api_url_builder import api_url_for, NODES, FILES, USERS
-from osfoffline.polling_osf_manager.remote_objects import RemoteFile, RemoteFolder, RemoteFileFolder, dict_to_remote_object, RemoteUser, RemoteNode
-osf_path = '/Users/himanshu/Desktop/OSF/'
-base_project_name = 'new_test_project'
-project_path = os.path.join(osf_path, base_project_name)
-USER_ID = 'k92s7' # SET ME
-
-# headers = {'Authorization':'Bearer {}'.format(USER_ID)}
-headers={'Cookie':'osf_staging=55fc5f29029bdb53541b5cda.wTLtvhA3IyD-UGpB3pr7YXIWHvc;'}
-session = requests.Session()
-session.headers.update(headers)
-files_dir_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'files')
-def create_new_node(title, parent=None):
-    #todo: relate to parent node
-    body = {
-            "data": {
-                "type": "nodes", # required
-                "attributes": {
-                    "title":    title,         # required
-                    "category": 'Project',      # required
-                }
-            }
-    }
-
-    headers['Content-Type']='application/json'
-    headers['Accept']='application/json'
-    headers['Cookie']='osf_staging=55fc5f29029bdb53541b5cda.wTLtvhA3IyD-UGpB3pr7YXIWHvc;'
-    ret = requests.post(api_url_for(NODES), data=json.dumps(body), headers=headers)
-    return RemoteNode(ret.json()['data']).id
-
-nid1 = '4e6k8' #create_new_node(base_project_name)
-
-@decorator
-def repeat_until_success(func, *args, **kwargs):
-    for i in range(20):
-        if func(*args, **kwargs):
-            return
-        else:
-            time.sleep(5)
-    raise TestFail
-
-
-def create_local(*args, file_name=None):
-    path = build_path(*args)
-    if not os.path.exists(path):
-        os.makedirs(path)
-    if file_name:
-        file_path = os.path.join(path, file_name)
-        file = open(file_path, 'w+')
-        contents = 'some text inside file {}'.format(file_path)
-        file.write(contents)
-        file.close()
-        return contents.encode('utf-8')
-
-
-
-def build_path(*args):
-    return os.path.join(project_path, *args)
 
 # usage: nosetests /path/to/manipulate_osf.py -x
 
-class TestFail(Exception):
-    pass
 
-
-def get_node_file_folders(node_id):
-    node_files_url = api_url_for(NODES, related_type=FILES, node_id=node_id)
-    resp = session.get(node_files_url)
-
-    assert resp.ok
-    osf_storage_folder = RemoteFolder(resp.json()['data'][0])
-    assert osf_storage_folder.provider == osf_storage_folder.name
-    children_resp = session.get(osf_storage_folder.child_files_url)
-    assert children_resp.ok
-    return [dict_to_remote_object(file_folder) for file_folder in children_resp.json()['data']]
-
-def get_children_file_folders(parent_folder):
-    assert isinstance(parent_folder, RemoteFolder)
-    url = parent_folder.child_files_url
-    resp = session.get(url)
-    assert resp.ok
-    return [dict_to_remote_object(file_folder) for file_folder in resp.json()['data']]
-
-def in_list(name, remote_object_list, is_dir):
-    cls = RemoteFolder if is_dir else RemoteFile
-    for remote_object in remote_object_list:
-        if isinstance(remote_object, cls) and remote_object.name == name:
-            return True
-    return False
-
-def get_from_list(name, remote_object_list, is_dir):
-    cls = RemoteFolder if is_dir else RemoteFile
-    for remote_object in remote_object_list:
-        if isinstance(remote_object, cls) and remote_object.name == name:
-            return remote_object
-    assert FileNotFoundError
-
-@repeat_until_success
-def assert_contains_folder(name, nid, parent_folder=None):
-    if parent_folder:
-        children = get_children_file_folders(parent_folder)
-    else:
-        children = get_node_file_folders(nid)
-
-    return in_list(name, children, True)
-
-@repeat_until_success
-def assert_contains_file(name, contents, nid, parent_folder=None):
-    if parent_folder:
-        children = get_children_file_folders(parent_folder)
-    else:
-        children = get_node_file_folders(nid)
-
-    if in_list(name, children, False):
-        file = get_from_list(name, children, False)
-        resp = session.get(file.download_url)
-        assert resp.ok
-        assert resp.content == contents
-        return True
-    return False
-
-
-@repeat_until_success
-def assert_file_not_exist(name, nid, parent_folder=None):
-    if parent_folder:
-        children = get_children_file_folders(parent_folder)
-    else:
-        children = get_node_file_folders(nid)
-
-    return not in_list(name, children, False)
-
-@repeat_until_success
-def assert_folder_not_exist(name, nid, parent_folder=None):
-    if parent_folder:
-        children = get_children_file_folders(parent_folder)
-    else:
-        children = get_node_file_folders(nid)
-    return not in_list(name, children, True)
-
-
-def get_remote(name, nid, is_dir, parent=None):
-    if parent:
-        children = get_children_file_folders(parent)
-    else:
-        children = get_node_file_folders(nid)
-    cls = RemoteFolder if is_dir else RemoteFile
-    for child in children:
-        if isinstance(child, cls) and child.name==name:
-            return child
-    raise FileNotFoundError
+# This code does NOT the OSFClient and AuthClient.
+# These tests are meant to be standalone
 
 
 
-def delete_all_local():
-    for file_folder in os.listdir(project_path):
-        path = os.path.join(project_path, file_folder)
-        if os.path.isdir(path):
-            shutil.rmtree(path)
-        else:
-            os.remove(path)
-
-@repeat_until_success
-def assert_node_has_no_file_folders(nid):
-    file_folders = get_node_file_folders(nid)
-    return len(file_folders)==0
-
-@repeat_until_success
-def assert_local_has_components_folder():
-    return os.path.isdir(build_path('Components'))
-
-def assert_contains_project():
-    if not os.path.isdir(project_path):
-        assert TestFail
 
 def setUp():
     assert_contains_project()
