+++ tests/acceptance/settings.py
@@ -1,0 +1,14 @@
+import os
+
+
+osf_path = '/path/to/OSF'
+base_project_name = 'sample_project'
+project_path = os.path.join(osf_path, base_project_name)
+files_dir_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'files')
+
+osf_user_id = 'xxxxx'
+
+
+
+
+
